third_party/blink/web_tests/fast/html/imports/import-same-url-expected.txt

CONSOLE WARNING: HTML Imports is deprecated and will be removed in M80, around February 2020. Please use ES modules instead. See https://www.chromestatus.com/features/5144752345317376 and https://developers.google.com/web/updates/2019/07/web-components-time-to-upgrade for more details.
PASS target1.import is non-null.
PASS target1.import is target2.import
PASS target1.import is target3.import
PASS successfullyParsed is true

TEST COMPLETE
